Ajout de la suppression d'un wiki Nettoyage des fichiers temporaires lors de l'archivages

+++ admin/index.php
@@ -7,34 +7,36 @@ include_once('php/view.class.php');
 $ferme = new Ferme("../ferme.config.php");
 $view  = new View($ferme);
 
-switch ($_GET['action']) {
-	case 'delete':
-		if(isset($_GET['name'])){
-			$ferme->delete($_GET['name']);
-			$view->addAlert("Wiki ".$_GET['name']." : Supprimé avec succès");
-			header("Location: ".$ferme->getURL());
-			exit;
-		}
-		break;
-
-	case 'save':
-		if(isset($_GET['name']))
-			try {
-				$ferme->save($_GET['name']);
-			} catch (Exception $e) {
-				$view->addAlert($e->getMessage(),"error");
+if(isset($_GET['action'])){
+	switch ($_GET['action']) {
+		case 'delete':
+			if(isset($_GET['name'])){
+				$ferme->delete($_GET['name']);
+				$view->addAlert("Wiki ".$_GET['name']." : Supprimé avec succès");
 				header("Location: ".$ferme->getURL());
 				exit;
 			}
-		
-			$view->addAlert("Wiki ".$_GET['name']." : Sauvegardé avec succès");
-			header("Location: ".$ferme->getURL());
-			exit;
-		break;
+			break;
 
-	default:
-		# rien, vraiment...
-		break;
+		case 'save':
+			if(isset($_GET['name']))
+				try {
+					$ferme->save($_GET['name']);
+				} catch (Exception $e) {
+					$view->addAlert($e->getMessage(),"error");
+					header("Location: ".$ferme->getURL());
+					exit;
+				}
+			
+				$view->addAlert("Wiki ".$_GET['name']." : Sauvegardé avec succès");
+				header("Location: ".$ferme->getURL());
+				exit;
+			break;
+
+		default:
+			# rien, vraiment...
+			break;
+	}
 }
 
 $view->show();
